Image Processing/requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0
tkinterdnd2>=0.4.2
matplotlib>=3.7.0
scikit-image>=0.21.0
tkinterdnd2
pyinstaller